requirements.txt

# Instead of install package normally with "pip install <package>" you could use "venv" for copying my configuration.
# For seeing all your packages, use "pip freeze".
# For use this requirements.txt file, use "python3 -m pip install -r requirements.txt"
# 
# refer to: https://packaging.python.org/en/latest/guides/installing-using-pip-and-virtual-environments/#using-requirements-files
#
# All dependencies
certifi==2020.6.20
charset-normalizer==3.1.0
idna==3.3
prometheus-client==0.16.0
pytz==2022.1
requests==2.28.2
urllib3==1.26.5
websocket-client==1.5.1
PyJWT==2.3.0
PyYAML==5.4.1
